requirements.txt
numpy<2
pydantic
scipy
--index-url https://pypi.org/simple/
reforge-core
xarm-python-sdk
git+https://github.com/twtoner/pimu.git